huyen_crm/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
click==8.0.4
Django==4.0.3
docx2pdf==0.1.8
Flask==2.0.3
image==1.5.33
itsdangerous==2.1.1
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.1
numpy==1.22.3
opencv-python==4.5.5.64
path==16.4.0
pdf2image==1.16.0
Pillow==9.0.1
PyPDF2==1.26.0
python-docx==0.8.11
six==1.16.0
sqlparse==0.4.2
tqdm==4.63.0
Unipath==1.1
Wand==0.6.7
Werkzeug==2.0.3
